simplify loading mechanism: remove _hooks.so
